% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2fst.R
\name{csv2fst}
\alias{csv2fst}
\title{Convert CSV files to FST format for better performance}
\usage{
csv2fst(csv_path, chunk_size = 50000)
}
\arguments{
\item{csv_path}{path to CSV or FST file}
}
\description{
Convert CSV files to FST format for better performance
}
