% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadApp.R
\name{downloadApp}
\alias{downloadApp}
\alias{downloadServer}
\alias{downloadInput}
\title{Download App Module}
\usage{
downloadApp()

downloadServer(id, main_par, peak_table, scan_table, plot_result)

downloadInput(id, main_par, import)
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory and annotation_list}

\item{trait_type}{character string as reactive object}
}
\description{
Download App Module
}
