% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadScanApp.R
\name{downloadScanApp}
\alias{downloadScanApp}
\alias{downloadScanServer}
\alias{downloadScanInput}
\title{Download Scan App Module}
\usage{
downloadScanApp()

downloadScanServer(id, main_par, peak_table, scan_table, plot_result)

downloadScanInput(id, main_par, import)
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory and annotation_list}

\item{trait_type}{character string as reactive object}
}
\description{
This is an adaptation from Kalynn code to download the QTL plot.
Plan is to use downloaddApp from foundrShiny package instead.
}
