% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_rows.R
\name{fst_rows}
\alias{fst_rows}
\title{Create row indices for FST files to speed up data access}
\usage{
fst_rows(fst_path)
}
\arguments{
\item{fst_path}{path to FST file}
}
\description{
Create row indices for FST files to speed up data access
}
