% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_alleles.R
\name{ggplot_alleles}
\alias{ggplot_alleles}
\title{Plot the Alleles}
\usage{
ggplot_alleles(peak, colors = newClrs)
}
\arguments{
\item{peak}{object with data to plot}

\item{colors}{character vector of colors for the plot}
}
\description{
Plot the Alleles
}
