% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_qtl_scan.R
\name{ggplot_qtl_scan}
\alias{ggplot_qtl_scan}
\title{Plot the QTL if doing a new scan}
\usage{
ggplot_qtl_scan(qtl_plot_obj, LOD_thr = NULL, xvar = "BPcum")
}
\arguments{
\item{qtl_plot_obj}{object with data to plot}

\item{LOD_thr}{threshold for LOD score}

\item{xvar}{variable for x-axis (default: "BPcum")}
}
\description{
Plot the QTL if doing a new scan
}
