% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly_qtl_scan.R
\name{ggplotly_qtl_scan}
\alias{ggplotly_qtl_scan}
\title{Plot the QTL if doing a new scan}
\usage{
ggplotly_qtl_scan(
  scan_table,
  peak_table,
  selected_chr = "2",
  source = "scanly_plot",
  plot_width = 900,
  plot_height = 600
)
}
\arguments{
\item{scan_table}{data frame with scan data}

\item{peak_table}{data frame with peak information}

\item{selected_chr}{selected chromosome}

\item{source}{name of the plotly source for click events}

\item{plot_width}{width of the plot}

\item{plot_height}{height of the plot}
}
\description{
Plot the QTL if doing a new scan
}
