% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data from files}
\usage{
import_data()
}
\description{
This function imports data from files specified in the `import.csv` file
located in the package's `data` directory. The function reads files with
extensions `.csv`, `.rds`, or `.xlsx` and returns a list of data frames.
}
