% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importApp.R
\name{importApp}
\alias{importApp}
\alias{importServer}
\alias{importUI}
\alias{importOutput}
\title{Import Data App Module}
\usage{
importApp()

importServer(id)

importUI(id)

importOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}

\item{annotation_list}{list with annotation information}

\item{set_type}{character string as reactive object}
}
\description{
Import Data App Module
}
