% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainParApp.R
\name{mainParApp}
\alias{mainParApp}
\alias{mainParServer}
\alias{mainParInput}
\alias{mainParUI}
\alias{mainParOutput}
\title{Main Parameter App Module}
\usage{
mainParApp(id)

mainParServer(id, import)

mainParInput(id)

mainParUI(id)

mainParOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{data}{reactive list with file_directory and annotation_list}
}
\description{
Main Parameter App Module
}
