% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainParServer.R
\name{mainParServer}
\alias{mainParServer}
\alias{mainParApp}
\title{Main parameter module}
\usage{
mainParServer(id, file_directory, annotation_list)

mainParApp(id)
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}

\item{annotation_list}{list with annotation information}

\item{set_type}{reactive object}
}
\description{
Main parameter module
}
