% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeApp.R
\name{mergeApp}
\alias{mergeApp}
\alias{mergeServer}
\title{Merge App}
\usage{
mergeApp(id)

mergeServer(id, primary_list, secondary_list)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{prefix}{static prefix for filename}

\item{main_par}{input parameters from calling routine}

\item{download_list}{reactiveValues with  postfix,plot,table}
}
\value{
nothing
}
\description{
Merge App
}
