% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakApp.R
\name{peakApp}
\alias{peakApp}
\alias{peakServer}
\alias{peakInput}
\alias{peakOutput}
\title{Peak App Module}
\usage{
peakApp()

peakServer(id, main_par, file_directory)

peakInput(id)

peakOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}
}
\description{
Peak App Module
}
