% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakServer.R
\name{peakServer}
\alias{peakServer}
\alias{peakApp}
\title{Peak module}
\usage{
peakServer(id, main_par, file_directory)

peakApp()
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}
}
\description{
Peak module
}
