% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_plot_visualizer.R
\name{QTL_plot_visualizer}
\alias{QTL_plot_visualizer}
\title{Plot the QTL if doing a new scan}
\usage{
QTL_plot_visualizer(qtl.temp, phenotype, LOD_thr, mrkrs)
}
\arguments{
\item{qtl.temp}{data frame of QTL scans}

\item{phenotype}{character string}

\item{LOD_thr}{value of LOD threshold}

\item{mrkrs}{list of markers}
}
\description{
Plot the QTL if doing a new scan
}
