% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlApp.R
\name{qtlApp}
\alias{qtlApp}
\alias{qtlServer}
\alias{qtlInput}
\alias{qtlOutput}
\title{QTL App}
\usage{
qtlApp()

qtlServer(id, file_directory, annotation_list, markers)

qtlInput(id)

qtlOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}

\item{annotation_list}{list with annotation information}

\item{markers}{list object with marker information}
}
\description{
QTL App
}
