% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlServer.R
\name{qtlServer}
\alias{qtlServer}
\alias{qtlOutput}
\alias{qtlApp}
\title{QTL Server}
\usage{
qtlServer(id, file_directory, annotation_list, markers)

qtlOutput(id)

qtlApp()
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}

\item{annotation_list}{list with annotation information}

\item{markers}{list object with marker information}
}
\description{
QTL Server
}
