% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanApp.R
\name{scanApp}
\alias{scanApp}
\alias{scanServer}
\alias{scanInput}
\alias{scanUI}
\alias{scanOutput}
\title{Scan App Module}
\usage{
scanApp()

scanServer(id, main_par, file_directory, annotation_list, markers)

scanInput(id)

scanUI(id)

scanOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}

\item{annotation_list}{list with annotation information}

\item{markers}{list object with marker information}
}
\description{
Scan App Module
}
