% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanlyApp.R
\name{scanlyApp}
\alias{scanlyApp}
\title{Plotly Scan App Module}
\usage{
scanlyApp()
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory and markers}

\item{main_par}{reactive list with selected_dataset, LOD_thr and which_trait}
}
\description{
Plotly Scan App Module
}
