% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanServer.R
\name{scanServer}
\alias{scanServer}
\title{Scan module}
\usage{
scanServer(id, main_par, file_directory, annotation_list, markers)
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}

\item{annotation_list}{list with annotation information}

\item{markers}{list object with marker information}
}
\description{
Scan module
}
