% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_scan.R
\name{trait_scan}
\alias{trait_scan}
\title{Find trait scan}
\usage{
trait_scan(file_dir, selected_dataset, selected_trait)
}
\arguments{
\item{file_dir}{data frame with file directory information}

\item{selected_dataset}{character string}

\item{selected_trait}{character string}
}
\description{
Find trait scan
}
