% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitApp.R
\name{traitApp}
\alias{traitApp}
\alias{traitServer}
\alias{traitUI}
\alias{traitOutput}
\title{Trait App Module}
\usage{
traitApp()

traitServer(id, main_par, import)

traitUI(id)

traitOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory and annotation_list}

\item{trait_type}{character string as reactive object}
}
\description{
Trait App Module
}
