% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitServer.R
\name{traitServer}
\alias{traitServer}
\alias{traitApp}
\title{Trait module}
\usage{
traitServer(id, file_directory, annotation_list, set_type)

traitApp()
}
\arguments{
\item{id}{shiny identifier}

\item{file_directory}{data frame with file directory information}

\item{annotation_list}{list with annotation information}

\item{set_type}{character string as reactive object}
}
\description{
Trait module
}
