% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitTypeApp.R
\name{traitTypeApp}
\alias{traitTypeApp}
\alias{traitTypeServer}
\title{Trait Type App Module}
\usage{
traitTypeApp()

traitTypeServer(id, main_par, import)
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory and annotation_list}

\item{trait_type}{character string as reactive object}
}
\description{
Trait Type App Module
}
