% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_qtl_scan.R
\name{ggplot_qtl_scan}
\alias{ggplot_qtl_scan}
\title{Plot the QTL if doing a new scan}
\usage{
ggplot_qtl_scan(scan_table, LOD_thr = NULL, selected_chr = "All")
}
\arguments{
\item{scan_table}{dataframe with data to plot}

\item{LOD_thr}{threshold for LOD score}

\item{selected_chr}{chromosome to plot (default is "All")}
}
\description{
Plot the QTL if doing a new scan
}
