% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly_qtl_scan.R
\name{ggplotly_qtl_scan}
\alias{ggplotly_qtl_scan}
\title{Plot the QTL if doing a new scan}
\usage{
ggplotly_qtl_scan(
  scan_plot,
  selected_chr = "2",
  plot_width = 900,
  plot_height = 600
)
}
\arguments{
\item{scan_plot}{ggplot object}

\item{selected_chr}{selected chromosome}

\item{plot_width}{width of the plot}

\item{plot_height}{height of the plot}
}
\description{
Plot the QTL if doing a new scan
}
