% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakApp.R
\name{peakApp}
\alias{peakApp}
\alias{peakServer}
\alias{peakInput}
\alias{peakUI}
\alias{peakOutput}
\title{Peak App Module}
\usage{
peakApp()

peakServer(id, main_par, import)

peakInput(id)

peakUI(id)

peakOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory}
}
\description{
Peak App Module
}
