% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlApp.R
\name{qtlApp}
\alias{qtlApp}
\alias{qtlServer}
\alias{qtlInput}
\alias{qtlOutput}
\title{QTL App}
\usage{
qtlApp()

qtlServer(id)

qtlInput(id)

qtlOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory, annotation_list and markers}
}
\description{
QTL App
}
