% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanApp.R, R/scanlyApp.R
\name{scanApp}
\alias{scanApp}
\alias{scanServer}
\alias{scanUI}
\alias{scanOutput}
\alias{scanlyServer}
\alias{scanlyUI}
\alias{scanlyOutput}
\title{Scan App Module}
\usage{
scanApp()

scanServer(id, main_par, import)

scanUI(id)

scanOutput(id)

scanlyServer(id, main_par, scan_list, peak_list)

scanlyUI(id)

scanlyOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{main_par}{reactive list with selected_dataset, LOD_thr and which_trait}

\item{import}{reactive list with file_directory and markers}
}
\description{
Scan App Module
}
