% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanApp.R
\name{scanApp}
\alias{scanApp}
\alias{scanServer}
\alias{scanInput}
\alias{scanUI}
\alias{scanOutput}
\title{Scan App Module}
\usage{
scanApp()

scanServer(id, main_par, import)

scanInput(id)

scanUI(id)

scanOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{import}{reactive list with file_directory and markers}
}
\description{
Scan App Module
}
