% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanlyApp.R
\name{scanlyApp}
\alias{scanlyApp}
\title{Plotly Scan App Module}
\usage{
scanlyApp()
}
\arguments{
\item{id}{shiny identifier}

\item{main_par}{reactive list with selected_dataset, LOD_thr and which_trait}

\item{scan_list}{reactive list from scanServer}

\item{peak_table}{reactive dataframe from peakServer}
}
\description{
This app has a point-click feature not yet implemented.
See `kalynn_R/latest_app_kalynn/app.R`.
Also not sure if `selected_chr` should be here or elsewhere.
}
